Add H2 db and .classpath watch out!
So H2 is nice and simple so far! I am using it a bit willy nilly but
that doesn't seem too bad for right now. I added the .classpath so that
it would error for anyone trying to build this somewhere else and
hopefully remind them to download h2 jar for themselves and change
.classpath. This should be changed some time to use maven probably but
not an undertaking for right now